requirements.txt
fastapi
uvicorn
httpx
kafka-python
